% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_garch.R
\name{auto_garch}
\alias{auto_garch}
\title{Fit the Best GARCH Model to an Univariate Time Series}
\usage{
auto_garch(R, variance = c("sGARCH", "eGARCH", "gjrGARCH", "apARCH",
  "csGARCH"), distributions = c("norm", "std", "ged", "snorm", "sstd",
  "sged", "jsu", "ghyp"), garch_p = c(0, 1), garch_q = c(0, 1),
  arma_p = c(0, 1), arma_q = c(0, 1), criteria = c("Akaike", "Bayes",
  "Shibata", "Hannan-Quinn", "likelihood"), n.ahead = 1,
  conditional = TRUE, ...)
}
\arguments{
\item{R}{A vector, matrix, data.frame, xts, timeSeries, zoo or a tibble object.}

\item{variance}{A vector or a list of character strings with the variance models to be computed. It can be any combination of: "sGARCH", "eGARCH", "gjrGARCH", "apARCH", "csGARCH".}

\item{distributions}{A vector or a list of character strings with the variance models to be computed. It can be any combination of: "norm", "std", "ged", "snorm", "sstd", "sged", "jsu", "ghyp".}

\item{garch_p, garch_q}{A vector or list with the number of mimimum and maximum number of lags to be included in the garch process.}

\item{arma_p, arma_q}{A vector or list with the mimimum and maximum number of lags to be included in the arma process.}

\item{criteria}{The criteria in which the models will be evaluated. One of: "Akaike", "Bayes", "Shibata", "Hannan-Quinn" and "likelihood".}

\item{n.ahead}{The number of periods ahead from which the sigmas should be forecasted.}

\item{conditional}{TRUE or FALSE. If TRUE, the the conditional sigmas covariances is returned. If FALSE, the unconditional covariance is printed.}

\item{...}{Any other parameters to pass thought ugarchspec.}
}
\value{
A variance-covariance matrix.
}
\description{
This function searches over different model specifications to find the best according to one of the selection criterias: Akaike, Bayes, shibata, Hannan-Quinn and likelihood.
}
\details{
This function searchs thought the best GARCH models by "brute force". Since fitting a Garch's can be demanding, it is advisable to not explore all the options for portfolio optimization unless you have time and a enough memory into your computer.
}
