% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_ewma.R
\name{cov_ewma}
\alias{cov_ewma}
\title{Exponentially Weighted Moving Average (EWMA) Covariance Matrix}
\usage{
cov_ewma(R, lambda = 0.96)
}
\arguments{
\item{R}{A vector, matrix, data.frame, xts, timeSeries or zoo object to be checked and coerced.}

\item{lambda}{A number between 0 and 1.}
}
\value{
A variance-covariance matrix.
}
\description{
This functions implements the covariance matrix as described in Longerstaey & Spencer (1996).
}
\examples{
cov_ewma(EuStockMarkets)
}
\references{
Longerstaey, Jacques, and Martin Spencer. \emph{Riskmetrics technical document.} Morgan Guaranty Trust Company of New York: New York 51 (1996): 54.
}
\author{
Bernardo Reckziegel
}
