% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_shrink_to_bayes_stein.R
\name{cov_shrink_to_bayes_stein}
\alias{cov_shrink_to_bayes_stein}
\title{Shrink the Variance-Covariance Matrix towards the Bayes-Stein Estimator}
\usage{
cov_shrink_to_bayes_stein(R)
}
\arguments{
\item{R}{A vector, matrix, data.frame, xts, timeSeries, zoo or a tibble object.}
}
\value{
A variance-covariance matrix.
}
\description{
Implements the variance-covariance matrix as described in Jorrion (1986),
}
\examples{
cov_shrink_bayes_stein(EuStockMarkets)
}
\references{
Jorion, Philippe. \emph{Bayes-Stein Estimation For Portfolio Analysis}. Journal of Financial and Quantitative Analysis 21.3 (1986): 279-292.
}
\author{
Bernardo Reckziegel
}
