% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_shrink_to_honey.R
\name{cov_shrink_to_honey}
\alias{cov_shrink_to_honey}
\title{Shrink the Covariance Matrix towards Hollywood!}
\usage{
cov_shrink_to_honey(R, shrink = NULL, portfolio = TRUE)
}
\arguments{
\item{R}{A vector, matrix, data.frame, xts, timeSeries or zoo object to be checked and coerced.}

\item{shrink}{A number between 0 and 1.}

\item{portfolio}{If \code{TRUE} only the covariance matrix is printed. If \code{FALSE} a list is returned with the covariance-matrix and the shrinkage intensity.}
}
\value{
A variance-covariance matrix.
}
\description{
Implements the covariance matrix described in Ledoit & Wold (2003).
}
\examples{
library(PortfolioAnalytics)
data(edhec)
cov_shrink_to_honey(edhec)
cov_shrink_to_honey(edhec, portfolio = FALSE)
}
\references{
Ledoit, Olivier, and Michael Wolf. \emph{Honey, I shrunk the sample covariance matrix.} UPF economics and business working paper 691 (2003).
}
\author{
Bernardo Reckziegel
}
