% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_shrink_to_identity.R
\name{cov_shrink_to_identity}
\alias{cov_shrink_to_identity}
\title{Shrink the Covariance Matrix towards to Identity}
\usage{
cov_shrink_to_identity(R, shrink = NULL, portfolio = TRUE)
}
\arguments{
\item{R}{A vector, matrix, data.frame, xts, timeSeries or zoo object to be checked and coerced.}

\item{shrink}{A number between 0 and 1.}

\item{portfolio}{If \code{TRUE} only the covariance matrix is printed. If \code{FALSE} a list is returned with the covariance-matrix and the shrinkage intensity.}
}
\value{
A variance-covariance matrix.
}
\description{
Implements the covariance matrix as described in Ledoit & Wolf (2004).
}
\examples{
cov_shrink_to_identity(EuStockMarkets)
cov_shrink_to_identity(EuStockMarkets, portfolio = FALSE)
}
\references{
Ledoit, Olivier, and Michael Wolf. \emph{A well-conditioned estimator for large-dimensional covariance matrices.} Journal of multivariate analysis 88.2 (2004): 365-411.
}
\author{
Bernardo Reckziegel
}
