% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_shrink_to_market.R
\name{cov_shrink_to_market}
\alias{cov_shrink_to_market}
\title{Shrink The Covariance Matrix towards to the Market Factor}
\usage{
cov_shrink_to_market(R, shrink = TRUE, portfolio = TRUE)
}
\arguments{
\item{R}{A vector, matrix, data.frame, xts, timeSeries or zoo object to be checked and coerced.}

\item{shrink}{A number between 0 and 1.}

\item{portfolio}{If \code{TRUE} only the covariance matrix is printed. If \code{FALSE} a list is returned with the covariance-matrix and the shrinkage intensity.}
}
\value{
A variance-covariance matrix.
}
\description{
Implements the covariance matrix as described in Ledoit & Wolf (2003).
}
\examples{
cov_shrink_to_market(EuStockMarkets)
}
\references{
Ledoit, Olivier, and Michael Wolf. \emph{Improved estimation of the covariance matrix of stock returns with an application to portfolio selection.} Journal of empirical finance 10.5 (2003): 603-621.
}
\author{
Bernardo Reckziegel
}
