% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_ewma.R
\name{mu_ewma}
\alias{mu_ewma}
\title{Exponentially Weighted Moving Average (EWMA) for Returns}
\usage{
mu_ewma(R, lambda = 0.96)
}
\arguments{
\item{R}{A vector, matrix, data.frame, xts, timeSeries or zoo object to be checked and coerced.}

\item{lambda}{A number between 0 and 1.}
}
\value{
A vector of assets expected returns.
}
\description{
This function implements EWMA for a vector of assets returns.
}
\examples{
mu_ewma(EuStockMarkets, lambda = 0.96)
}
\references{
Longerstaey, Jacques, and Martin Spencer. \emph{Riskmetrics technical document.} Morgan Guaranty Trust Company of New York: New York 51 (1996): 54.
}
\author{
Bernardo Reckziegel
}
