% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_moments.R
\name{portfolio_moments}
\alias{portfolio_moments}
\title{Specify custom functions to be used with Portfolio Analytics}
\usage{
portfolio_moments(R, mu, sigma)
}
\arguments{
\item{R}{A vector, matrix, data.frame, xts, timeSeries, zoo or a tibble object.}

\item{mu}{A function, formula, or vector (not necessarily atomic). Follows the same syntax as the \code{\link[purrr:map]{map}} family of functions.}

\item{sigma}{A function, formula, or vector (not necessarily atomic). Follows the same syntax as the \code{\link[purrr:map]{map}} family of functions.}
}
\value{
A list with two components: \code{mu} and \code{sigma}.
}
\description{
\code{portfolio_moments} uses \code{purrr} syntax for formulas. This allows a flexible and succinct way to define anonymous (lambda) functions.
}
\details{
\code{mu} and \code{sigma} can be a passed as a function, a formula or a vector:

\itemize{
     \item If a __function__: use it as it is.
     \item If a __formula__: e.g. \code{~ . + 2} is converted to a function. If the function requires a time series to run refer to it with \code{.}.
     \item If a __vector__ e.g. \code{~ 0.01} is converted to a function and all forecasts get the same value, 0.01.
   }

A \code{R}, the first argument of \code{portfolio_moments}, does not need to be specified in \code{\link{optimize_portfolio}}. Use only \code{mu} and \code{sigma}.
}
\examples{
library(PortfolioAnalytics)
data(edhec)

# function format:
portfolio_moments(
    R     = edhec,
    mu    = auto.arima,
    sigma = cov_shrink_to_market
)

# formula format: use to customize your call
# (ps: works for "sigma" as well)
portfolio_moments(
    R     = edhec,
    mu    = ~ auto.arima(y = ., max.p = 2, max.q = 2, stationary = TRUE),
    sigma = cov_shrink_to_market
)

# vector format:
portfolio_moments(
    R     = edhec,
    mu    = ~ 0.006, # this estimator is 100\% bias and 0\% variance.
    sigma = cov_shrink_to_market
)


# When optimizing the portfolio use "mu" and "sigma" only.
portfolio.spec(assets = colnames(edhec)) \%>\%
 add.constraint(portfolio = ., type = "box", min = 0.00, max = 1.00) \%>\%
 add.objective(portfolio = ., type = "risk", name = "var") \%>\%
 add.objective(portfolio = ., type = "return", name = "mean") \%>\%
 optimize_portfolio(
    R               = edhec,
    portfolio       = mean_var_spec,
    optimize_method = 'ROI',
    search_size     = 500,
    mu              = ets  # no need to use portfolio_moments()
    sigma           = cov  # just "mu" and "sigma"
)
# Note that the sample-mean and the sample-covariance are also accepted.
}
\seealso{
[purrr::map()], [purrr::as_mapper()], [[rlang::as_function()] and [PortfolioAnalytics::optimize.portfolio()]
}
\author{
Bernardo Reckziegel
}
