% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_edhec.R
\name{update_edhec}
\alias{update_edhec}
\title{Update the EDHEC}
\usage{
update_edhec(class = NULL)
}
\arguments{
\item{class}{One of 'tbl', 'xts', 'zoo', 'zooreg' or 'ts'.}
}
\value{
An object of the same class required in the 'class' argument.
}
\description{
Download an updated version of the Edhec Headge Fund index.
}
\examples{
# xts object
head(update_edhec('xts'))

# tibble object
update_edhec('tbl')
}
