% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_portfolio_rebalancing.R
\name{optimize_portfolio_rebalancing}
\alias{optimize_portfolio_rebalancing}
\alias{optimize_portfolio_rebalancing.default}
\alias{optimize_portfolio_rebalancing.matrix}
\alias{optimize_portfolio_rebalancing.xts}
\alias{optimize_portfolio_rebalancing.zoo}
\alias{optimize_portfolio_rebalancing.data.frame}
\alias{optimize_portfolio_rebalancing.tbl_df}
\title{Optimize and Rebalance Portfolios Using Custom Functions}
\usage{
optimize_portfolio_rebalancing(R, ...)

\method{optimize_portfolio_rebalancing}{default}(R, ...)

\method{optimize_portfolio_rebalancing}{matrix}(R, ...)

\method{optimize_portfolio_rebalancing}{xts}(R, ...)

\method{optimize_portfolio_rebalancing}{zoo}(R, ...)

\method{optimize_portfolio_rebalancing}{data.frame}(R, ...)

\method{optimize_portfolio_rebalancing}{tbl_df}(R, select, date_var,
  silent = TRUE, ...)
}
\arguments{
\item{R}{An xts, matrix, data frame, timeSeries, zoo or tibble object of asset returns. The data cannot be "tidy".}

\item{...}{Any other paramenters to be passed into [PortfolioAnalytics::optimize.portfolio.rebalancing()].}
}
\value{
Same output of \code{optimize.portfolio.rebalancing}.
}
\description{
A wrapper around \code{optimize.portfolio.rebalancing()} that allows the user to quickly add forecasts for the mean and the variance.
}
\details{
As explained in the vignette "Custom Moment and Objective Functions", Portfolio Analytics works more efficiently when the mean is specified as _mu_ and the variance as _sigma_.
}
\examples{
# load data
data(edhec)

# create a mean-variance portfolio (long-only)
# that uses the \\code{auto.arima} function from the \\code{forecast} package as one step ahead forecasts for the mean
# and the famous "Bayes-Stein" estimator the covariance
mean_var_spec <- portfolio.spec(assets = colnames(edhec)) \%>\%
 add.constraint(portfolio = ., type = "box", min = 0.00, max = 1.00) \%>\%
 add.objective(portfolio = ., type = "risk", name = "var") \%>\%
 add.objective(portfolio = ., type = "return", name = "mean")

optimize_portfolio_rebalancing(
    R = edhec,
    portfolio = mean_var_spec,
    optimize_method = 'ROI',
    rebalance_on = "months",
    training_period = 60,
    rolling_window = 60,
    search_size = 500,
    mu = ets,
    sigma = shrink_bayes_stein
)


# \\code{PortfolioMoments} aloows you to use the same formulas syntax as \\code{purrr}.
# Suppose you want to add restrictions to the \\code{auto.arima} model above.

\dontrun{
optimize_portfolio(
    R = edhec,
    portfolio = mean_var_spec,
    optimize_method = 'ROI',
    rebalance_on = "months",
    training_period = 60,
    rolling_window = 60,
    search_size = 500,
    mu = ~ auto.arima(., max.p = 2, max.q = 2, seasonal = FALSE)
    sigma = shrink_bayes_stein #PortfolioMoments package
)
}
}
\seealso{
[PortfolioAnalytics::optimize.portfolio.rebalancing()] and [purrr::as_mapper()]
}
