% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_edhec.R
\name{update_edhec}
\alias{update_edhec}
\title{Title}
\usage{
update_edhec(class)
}
\arguments{
\item{class}{One of 'tbl', 'xts', 'zoo', 'zooreg' or 'ts'.}
}
\value{
An object of the same class required in the 'class' argument.
}
\description{
Download an updated version of the Edhec Headge Fund index.
}
\examples{
# xts object
update_edhec('xts') \%>\%
  head()

# tibble object
update_edhec('tbl')
}
